Bug 589151 – Let filechooser fall back to user's pictures folder
Add a (for now) hidden option to have the filechooser show the user's
picture folder ($XDG_PICTURES_DIR) when no image is loaded in the
current window. If the folder had not been set up it will fall back to
show the currrent working directory. The option has been enabled by
default as it seems to be useful on the average desktop system.