Add func to get the close confirmation dialog in an insensitive state
This state is special therein that it is not making the whole dialog
insensitive, but only the buttons and the checkbox column in the
treeview. That way the main window doesn't need to be made insensitive
during saving to avoid interaction. Alltogether this should improve the
situation visually quite a bit.