Reorder menu items to make menus more intuitive and less confusing
Based on a proposal by Hylke Bons. Most visible change is the 'File'
menu now being more specifically called as 'Image' now as that's what
we're talking about in this case.
Also the action names were adjusted to be in line with the others.
This causes custom toolbars to loose these items on the first start.
(bug 614457)