bgo#659830 - Arrow keys should always navigate, never scroll
We use plain arrow keys to navigate the image list, regardless of whether the image is
zoomed or whether it has the focus.  We use Alt+arrows to scroll the image.

This is to make it easier to browse an image collection while at the same time zooming
on images from time to time.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>