EogScrollView: add image rotation gesture/signal
A ::rotation-changed signal has been added to EogScrollView, that reports
angle changes happening from within the view in 90 degree steps. A
GtkGestureRotate has been hooked to the widget so two-finger touchscreen
interaction does rotate the image.

In order to avoid unintended angle changes when the delta is too close to
a switch point, there is a "grey area" threshold that makes it harder to
switch forth and back again.

This gesture has been added to the same group than the GtkGestureZoom,
as both are meant to share sequences and state.