EogWindow: remove the string matching from ui_settings_changed_cb
The changed setting is now passed as a detailed signal instead
of having to do string matching inside eog_window_ui_settings_changed_cb.
In order to facilitate this change, the connecting of the callbacks had
to be moved down a few lines to after the GActions are added to the window.

https://bugzilla.gnome.org/show_bug.cgi?id=741050