EogImage: Fix too early size-prepared emission
If an image without metadata support is loaded it could happen that
the size-prepared signal was emitted without knowing the actual
image size. That resulted in a smaller than necessary window size.
Remove unneeded emission since the loader will make sure to emit the
signal once the size is known.

https://bugzilla.gnome.org/show_bug.cgi?id=748598