EogWindow: Fix keyboard controls in the popover menu
This imports and adapts a workaround from gedit that keeps the keyboard
controls working in the popover menu and other widgets in the headerbar.
We lose some inofficial keyboard shortcuts in the process, but as their
names say, they were "inofficial". Maybe we can find a way to make them
work again in the future.

https://bugzilla.gnome.org/show_bug.cgi?id=754985