build: Remove unused variables and defines
Although micro and macro versions are commonly used, they are not
in Eye of GNOME, so there is no need to create them.

Some PACKAGE_* definitions are also not used in Eye of GNOME.

Finally, there is a variable pointing out if a debug build has been
enabled or not. This is used only once, so it's not necessary to
hold this information.

All of these have been removed.

https://bugzilla.gnome.org/show_bug.cgi?id=793719