build: Do not use options as auto
There are some options that can be used to enable or disable
options. However, if the conditions to use those components
enabled by options are not fulfilled, they are disabled and this
might confuse an user.

Now, if the conditions of those enabled components are not fulfilled
the build configuration will stop showing a message to the user.

meson variables have been renamed from the `have_` pattern to the
`enable_` pattern to better reflect this behaviour.

The `zlib` detection behaviour has also been fixed, which will
require both `inflate` and `crc32` to be available.

https://bugzilla.gnome.org/show_bug.cgi?id=793719