build: Reuse pkglibdir meson variable
The installation directory used on the build of the shared library
was used in the pkg-config file generation. Due to this behaviour
when using the full path as shared library generation caused the
libdir path to be wrong.

However, this has been fixed in meson's 0.44 version so the
pkglibdir directory is reused for this.

The required meson version has also been updated accordingly.

https://bugzilla.gnome.org/show_bug.cgi?id=793719