EogWindow: Fix skipping images in slideshows on slow devices
If a slow device is used for accessing files in a slideshow some images
get lost. This happens because sometimes images can not be loaded before
the slideshow switches to the next slide.
With this bugfix the timer does not tick in a fixed interval anymore.
The timer starts after the current image is successfully loaded.

Fixes #15