EogImage: Use 96dpi when rendering SVG documents
The CSS standard specifies 96dpi as a conversion factor for units
different than pixels:
https://www.w3.org/TR/css3-values/#absolute-lengths

This value should be assumed for SVG documents too.

Other programs (Firefox, Inkscape) also use 96dpi when converting units
of SVG files to pixels while rsvg, which EOG relies on, uses 90dpi. This
results in EOG rendering SVG documents with different dimensions
compared to said programs, when the original units are not in pixels.

Unfortunately changing the default value in librsvg itself is not viable
because that would break ABI compatibility, so fix the issue in EOG.

Fixes #22.