build: Define meson information early
The used meson modules, default directories and includes have been
moved to the start of the build file, just after project related
information, so they are available early.

The way some directory paths are defined has also been changed to
avoid the use of the `source_root` function because is discouraged.