EogWindow: Use GtkAppChooserDialog for the "Open With" choices
We want the list of applications to be available when Eog is
running inside of a Flatpak. For that, we should unify the UI/UX
of the "Open With" option.

With these changes, we open a GtkAppChooserDialog when the user
clicks in "Open With".

In a following patch we will make sure that the Flatpak Portal is
used when in a Flatpak environment, and that the dialog is used
otherwise.

Fixes #98