EogWindow: Avoid warning if collection is updated while reloading
The function to update the image position in the statusbar needs to
take into account that the window may not have an image loaded although
there are images in the collection. This can occur if the current image
is reloaded. Updating the position text would have spawned a critical
warning from eog_list_store_get_pos_by_image().