Add field `indices_changed`
When files are created or deleted, we can no longer rely on the
previous index range of start_thumb to end_thumb to be correct. This
flag gets set on creates and deletes and cleared when thumbs are
updated.

This is to address a crash caused by a rapid number of deletions which
left us in a state where end_thumb < start_thumb.

Fixes #189