build: Improve custom targets generation
Since meson 0.60, the name in `custom_target`, and derivatives, is
optional[0], although the implementation has produced some errors at
the moment[1].

Due to this, the name has been removed when possible.

When possible, two pass targets, where there is a first pass to
replace variables, and second pass to translate files, have been
modified to be avoid extra build commands.

The `@BASENAME@` token has also been used as an approach to remove
the `in` suffix.

[0] https://mesonbuild.com/Release-notes-for-0-60-0.html#optional-custom_target-name
[1] https://github.com/mesonbuild/meson/issues/9441