flatpak: Pin libportal dependency to release 0.4
We would need to update the dependency for changes in the build process,
however also the API is currently changing. Upstream has already
suggested in another case (epiphany!1027) to stay on a tagged release
until a new stable release is available. Let's do the same for eog.

This is a fixup of d74e06cea08a00725472dcaf7ca7d6e3a0707719 which I
accidentally pushed without updating after fixing the error.