EogWindow: Also set wallpaper as dark wallpaper
When we set a wallpaper from eog this configures the newly selected
wallpaper also for the new dark mode in GNOME 42 in addition to the
existing key that is used for the light mode. Otherwise the image
would not be used/updated. This is also the behaviour we get when
we change the wallpaper through the desktop portal.

The gsettings-desktop-schemas dependency needs bumping to at least 42.beta
to ensure the dark mode key is present.

Fixes #229.