EogWindow: Set window title additionally through GTK+ API
As HdyHeaderBar doesn't react to calls to `gtk_window_set_title` we
changed these calls to `hdy_header_bar_set_title` instead. But that
didn't forward these calls to the window manager which would only show
the application name for all windows.

This now calls both functions to keep the windows title in sync for both,
the headerbar and the window manager.

Fixes #244.