flatpak: Add webp pixbuf loader to the build
Since 1eb9734e2 eog assumes the webp loader to be present. However the
loader is not included in the Flatpak runtime yet, so we need to build
it ourselves.

This mirrors the build process that is used in the Flathub build that
creates a new cache file for gdk-pixbuf that enables the new loader and
keeps the loaders included with the runtime working.

Fixes #250.