EogJobs: Only emit "finished" signal if job wasn't cancelled
Emitting the "finished" signal happens delayed through the mainloop
while the job is executed async by the JobScheduler. In the time between
queuing the signal emission until it happens the job could be cancelled.
Do not emit the signal then as the caller is obviously no longer
interested in the job's result.

Part of #263.