EogListStore: Use`gtk_list_store_insert_with_values` to add images
This causes new images to be already at the correct position when the
"row-inserted" signal is emitted.

That should make the image index shown in the statusbar to be correct
when adding images in the middle. Previously the current image's index
would've been initially incorrect as it would count while the new image
was still at the end of the list.

Fixes #300.