Eog*: Use fast content type as fallback
It turns out that, depending on the responsible GVfs implementation,
a GFileInfo may not actually set the content type attribute even if
requested. Since knowing the content type is rather critical for eog
try to use the fast content type as a fallback in those cases.

The fast content type should be hardly unknown as it is usually just
based on the file extension.

Fixes #312.