EogListStore: Handle "moved-in" event like "file-changed"
This should avoid thumbnails getting out of sync when the temporary save
file can be directly `rename()`d to overwrite the target file. This
would trigger only a "moved-in" event in the file monitor which didn't
handle file changes yet, and thus also didn't refresh the thumbnail.

The behaviour while saving depends on how the distro sets up /tmp.
Distros using a tmpfs would not be affected by this, as GIO would switch
to a copy+rename cycle instead, only triggering a "rename" signal in the
directory monitor.

That should fix the issue analyzed by Robert Vidal in !151.
Drawback: Saving an image will now always trigger the messagebar asking
the user to reload the image. That however was already the case with the
"rename" events and thus is likely a seperate issue.